Clean up buckets, rename to files
